frontend work for dictionary
