+++ textdb/textdb-web/src/test/java/edu/uci/ics/textdb/web/resource/SystemResourceTest.java
@@ -6,6 +6,7 @@ import io.dropwizard.client.JerseyClientBuilder;
 import io.dropwizard.testing.ResourceHelpers;
 import io.dropwizard.testing.junit.DropwizardAppRule;
 import org.glassfish.jersey.client.ClientProperties;
+import org.glassfish.jersey.media.multipart.MultiPartFeature;
 import org.junit.ClassRule;
 import org.junit.Test;
 import javax.ws.rs.client.Client;
@@ -29,6 +30,37 @@ public class SystemResourceTest {
 
 		Response response = client.target(
 						String.format("http://localhost:%d/api/resources/metadata", RULE.getLocalPort()))
+						.request()
+						.get();
+
+		assertThat(response.getStatus()).isEqualTo(200);
+	}
+
+
+	@Test
+	public void getDictionaries() throws Exception {
+		Client client = new JerseyClientBuilder(RULE.getEnvironment()).build("test client");
+		client.property(ClientProperties.CONNECT_TIMEOUT, 5000);
+		client.property(ClientProperties.READ_TIMEOUT, 5000);
+		client.register(MultiPartFeature.class);
+
+		Response response = client.target(
+						String.format("http://localhost:%d/api/resources/dictionaries", RULE.getLocalPort()))
+						.request()
+						.get();
+
+		assertThat(response.getStatus()).isEqualTo(200);
+	}
+
+	@Test
+	public void getDictionaryPath() throws Exception {
+		Client client = new JerseyClientBuilder(RULE.getEnvironment()).build("test client");
+		client.property(ClientProperties.CONNECT_TIMEOUT, 5000);
+		client.property(ClientProperties.READ_TIMEOUT, 5000);
+		client.register(MultiPartFeature.class);
+
+		Response response = client.target(
+						String.format("http://localhost:%d/api/resources/dictionary/?id=tempid", RULE.getLocalPort()))
 						.request()
 						.get();
